popover: Correctly regain the device grab when broken
In our windows we would like to also support further child windows that
use grabs, i.e. popup menus for text entries and such. To do so we must
follow the chain of events for grabs, and when we become unshadowed by
a child grab, with our own device grab broken, attempt to restore that
grab once more.

This also allows us to keep the window as a "popup" type window, which
is more suitable for usage within a panel.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>