Implement the basic Y positioning of the popover
Instead of trying to be clever (and failing) by looking at the various
screen edges, we'll rely on the fact we're meant for use only in the
panel, and check what edge the toplevel panel actually lives on.

We're then able to determine the Y position from this. However our
tail code now needs fixing to handle the various positions, as well
as adding some clamp code for the complete screen bounds.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>