Add support for automatic placement of the popover
This uses a very simple bit of logic to determine where to place the
popover when in automatic mode. If we can fit the popover underneath
the current widget, then we'll always try to do so. If we can't, we'll
then check if we can put it on top of the widget. Failing this, we'll
then determine which side of the widget has more room, left or right,
and place the popover there. We can then rely on bounding and y
tail offsets to make it sync up nicely.

This fixes #1.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>