Implement full destruction chain for bound languages
If a widget dies, we remove it from the popover manager. Additionally,
a popover will check when it's associated widget dies too, and kill itself.
Thus, destruction of a widget leads to it being removed from the hashtable
of known popovers + widgets, and the popover itself is nuked.

This fixes #4.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>